test(HanScore): 한나라 보너스 점수 계산 책임 위임에 따른 테스트 수정
