+++ src/test/java/janggi/domain/board/BoardTest.java
@@ -89,16 +89,6 @@ class BoardTest {
     }
 
     @Test
-    @DisplayName("한나라의 기물이 차와 포만 있으면 총합은 1.5점 가산점을 더해서 21.5이다.")
-    void han_total_score() {
-        Map<Point, Piece> pieces = new LinkedHashMap<>();
-        pieces.put(Point.of(0, 0), new Cha(Team.HAN));
-        pieces.put(Point.of(1, 1), new Pho(Team.HAN));
-        board = new Board(pieces);
-        Assertions.assertThat(board.calculateScore(Team.HAN)).isEqualTo(21.5);
-    }
-
-    @Test
     @DisplayName("초나라의 기물이 차와 포만 있으면 총합은 20이다.")
     void cho_total_score() {
         Map<Point, Piece> pieces = new LinkedHashMap<>();
